Assistant checkpoint: Fix upload errors and missing hold probability imports
Assistant generated file changes:
- server/routes.ts: Add missing imports and fix hold probability recalculation, Fix API endpoint query handling and add seniorityPercentile to user creation
- client/src/lib/api.ts: Fix createOrUpdateUser to include seniorityPercentile, Fix searchPairings to use correct API endpoint
- client/src/components/profile-modal.tsx: Update profile modal to pass seniorityPercentile to API
- server/pdfParser.ts: Fix PDF parsing to work with seniority percentile updates, Update parsePDF method signature
- server/holdProbabilityCalculator.ts: Complete the missing parts of HoldProbabilityCalculator

---

User prompt:

the pairing were dropped out of the app and if i try to upload a new one, it fails saying failed to upload bid package, try again

+++ server/routes.ts
@@ -4,8 +4,9 @@ import { storage } from "./storage";
 import { seedDatabase } from "./seedData";
 import { pdfParser } from "./pdfParser";
 import { db } from "./db";
-import { eq } from "drizzle-orm";
+import { eq, gte, lte, sql } from "drizzle-orm";
 import { pairings } from "@shared/schema";
+import { HoldProbabilityCalculator } from "./holdProbabilityCalculator";
 
 import { openaiAssistant } from "./openaiAssistant";
 import multer from "multer";
@@ -345,7 +346,7 @@ export async function registerRoutes(app: Express): Promise<Server> {
 
       const user = await storage.createOrUpdateUser({
         seniorityNumber: parseInt(seniorityNumber),
-        seniorityPercentile: seniorityPercentile ? parseInt(seniorityPercentile) : 50,
+        seniorityPercentile: seniorityPercentile ? parseFloat(seniorityPercentile) : 50,
         base,
         aircraft
       });
